[update] rename gui.py to gui_display.py
